backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 数据处理
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF处理
pdfplumber==0.11.7

# 文件处理
aiofiles==23.2.1

# 工具库
python-dateutil==2.8.2
email-validator==2.1.0 